+++ termal_core/src/image/rect.rs
@@ -12,4 +12,9 @@ impl Rect {
     pub fn new(x: f32, y: f32, w: f32, h: f32) -> Self {
         Self { x, y, w, h }
     }
+
+    /// Gets the center of the rectangle.
+    pub fn center(&self) -> (f32, f32) {
+        (self.x + self.w / 2., self.y + self.h / 2.)
+    }
 }